+++ khal/ui/editor.py
@@ -20,16 +20,14 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 from datetime import datetime, time
+import datetime as dt
 
 import urwid
 
-from urwid import Edit
-
+from ..utils import get_weekday_occurrence
 from .widgets import DateWidget, TimeWidget, NColumns, NPile, ValidatedEdit, \
-    DateConversionError, Choice, PositiveIntEdit, AlarmsEditor, NListBox
+    DateConversionError, Choice, PositiveIntEdit, AlarmsEditor, NListBox, ExtendedEdit
 from .calendarwidget import CalendarWidget
-
-NOREPEAT = 'No'
 
 
 class StartEnd(object):
@@ -82,13 +80,18 @@ class CalendarPopUp(urwid.PopUpLauncher):
 
 
 class DateEdit(urwid.WidgetWrap):
+    """Widget that allows editing a Date.
+
+    Will open a calendar when `enter` is pressed, pressing `enter` again will
+    select that date.
+    """
+
     def __init__(
-            self,
-            startdt=None, dateformat='%Y-%m-%d', datewidth=10, caption='',
+            self, startdt=None, dateformat='%Y-%m-%d',
             on_date_change=lambda _: None,
-            weeknumbers=False, firstweekday=0,
-            keybindings=None,
+            weeknumbers=False, firstweekday=0, keybindings=None,
     ):
+        datewidth = len(startdt.strftime(dateformat)) + 1
         self._dateformat = dateformat
         if startdt is None:
             startdt = dt.date.today()
@@ -96,11 +99,10 @@ class DateEdit(urwid.WidgetWrap):
             dateformat=dateformat,
             EditWidget=DateWidget,
             validate=self._validate,
-            caption=('', caption),
             edit_text=startdt.strftime(dateformat),
             on_date_change=on_date_change)
         wrapped = CalendarPopUp(self._edit, on_date_change, weeknumbers, firstweekday, keybindings)
-        padded = urwid.Padding(wrapped, align='left', width=datewidth + 7, left=0, right=1)
+        padded = urwid.Padding(wrapped, align='left', width=datewidth, left=0, right=1)
         super().__init__(padded)
 
     def _validate(self, text):
@@ -113,31 +115,50 @@ class DateEdit(urwid.WidgetWrap):
 
     @property
     def date(self):
+        """Get currently entered date, or False, if input is invalid.
+
+        :returns: the currently entered date
+        :rtype: datetime.date
+        """
         return self._validate(self._edit.get_edit_text())
+
+    @date.setter
+    def date(self, date):
+        """Update text of this Widget.
+
+        :type date: datetime.date
+        """
+        self._edit.set_edit_text(date.strftime(self._dateformat))
 
 
 class StartEndEditor(urwid.WidgetWrap):
-    """Widget for editing start and end times (of an event)"""
+    """Widget for editing start and end times (of an event)."""
 
-    def __init__(self, start, end, conf, on_date_change=lambda x: None):
+    def __init__(self, start, end, conf,
+                 on_start_date_change=lambda x: None,
+                 on_end_date_change=lambda x: None,
+
+                 ):
         """
         :type start: datetime.datetime
         :type end: datetime.datetime
-        :param on_date_change: a callable that gets called everytime a new
-            date is entered, with that new date as an argument
+        :param on_start_date_change: a callable that gets called everytime a new
+            start date is entered, with that new date as an argument
+        :param on_end_date_change: same as for on_start_date_change, just for the
+            end date
         """
         self.allday = not isinstance(start, datetime)
         self.conf = conf
         self._startdt, self._original_start = start, start
         self._enddt, self._original_end = end, end
-        self.on_date_change = on_date_change
+        self.on_start_date_change = on_start_date_change
+        self.on_end_date_change = on_end_date_change
         self._datewidth = len(start.strftime(self.conf['locale']['longdateformat']))
         self._timewidth = len(start.strftime(self.conf['locale']['timeformat']))
         # this will contain the widgets for [start|end] [date|time]
         self.widgets = StartEnd(None, None, None, None)
-
-        self.checkallday = urwid.CheckBox('Allday', state=self.allday,
-                                          on_state_change=self.toggle)
+        self.checkallday = urwid.CheckBox(
+            'Allday', state=self.allday, on_state_change=self.toggle)
         self.toggle(None, self.allday)
 
     def keypress(self, size, key):
@@ -197,7 +218,7 @@ class StartEndEditor(urwid.WidgetWrap):
 
     def _start_date_change(self, date):
         self._startdt = self.localize_start(datetime.combine(date, self._start_time))
-        self.on_date_change(date)
+        self.on_start_date_change(date)
 
     def _validate_end_time(self, text):
         try:
@@ -210,7 +231,7 @@ class StartEndEditor(urwid.WidgetWrap):
 
     def _end_date_change(self, date):
         self._enddt = self.localize_end(datetime.combine(date, self._start_time))
-        self.on_date_change(date)
+        self.on_end_date_change(date)
 
     def toggle(self, checkbox, state):
         """change from allday to datetime event
@@ -231,16 +252,14 @@ class StartEndEditor(urwid.WidgetWrap):
             self._enddt = self._enddt.date()
         self.allday = state
         self.widgets.startdate = DateEdit(
-            self._startdt, self.conf['locale']['longdateformat'], self._datewidth, 'From: ',
-            self._start_date_change,
-            self.conf['locale']['weeknumbers'], self.conf['locale']['firstweekday'],
-            self.conf['keybindings'],
+            self._startdt, self.conf['locale']['longdateformat'],
+            self._start_date_change, self.conf['locale']['weeknumbers'],
+            self.conf['locale']['firstweekday'], self.conf['keybindings'],
         )
         self.widgets.enddate = DateEdit(
-            self._enddt, self.conf['locale']['longdateformat'], self._datewidth, 'To:   ',
-            self._end_date_change,
-            self.conf['locale']['weeknumbers'], self.conf['locale']['firstweekday'],
-            self.conf['keybindings'],
+            self._enddt, self.conf['locale']['longdateformat'],
+            self._end_date_change, self.conf['locale']['weeknumbers'],
+            self.conf['locale']['firstweekday'], self.conf['keybindings'],
         )
 
         if state is True:
@@ -271,10 +290,10 @@ class StartEndEditor(urwid.WidgetWrap):
 
         columns = NPile([
             self.checkallday,
-            NColumns([(self._datewidth + 7, self.widgets.startdate), (
+            NColumns([(5, urwid.Text('From:')), (self._datewidth, self.widgets.startdate), (
                 timewidth, self.widgets.starttime)], dividechars=1),
             NColumns(
-                [(self._datewidth + 7, self.widgets.enddate),
+                [(5, urwid.Text('To:')), (self._datewidth, self.widgets.enddate),
                  (timewidth, self.widgets.endtime)],
                 dividechars=1)
         ], focus_item=1)
@@ -301,7 +320,6 @@ class EventEditor(urwid.WidgetWrap):
         :param always_save: save event even if it has not changed
         :type always_save: bool
         """
-
         self.pane = pane
         self.event = event
         self._save_callback = save_callback
@@ -316,9 +334,14 @@ class EventEditor(urwid.WidgetWrap):
         self.categories = event.categories
         self.startendeditor = StartEndEditor(
             event.start_local, event.end_local, self._conf,
-            self.pane.eventscolumn.original_widget.set_focus_date)
-        self.recurrenceeditor = RecurrenceEditor(self.event.recurobject, self._conf, event.start_local)
-        self.summary = urwid.Edit(caption='Title: ', edit_text=event.summary)
+            self.start_datechange, self.end_datechange,
+        )
+        # TODO make sure recurrence rules cannot be edited if we only
+        # edit one instance (or this and future) (once we support that)
+        self.recurrenceeditor = RecurrenceEditor(
+            self.event.recurobject, self._conf, event.start_local,
+        )
+        self.summary = ExtendedEdit(caption='Title: ', edit_text=event.summary)
 
         divider = urwid.Divider(' ')
 
@@ -330,9 +353,11 @@ class EventEditor(urwid.WidgetWrap):
             self.collection._calendars[self.event.calendar],
             decorate_choice
         )
-        self.description = Edit(caption='Description: ', edit_text=self.description, multiline=True)
-        self.location = Edit(caption='Location: ', edit_text=self.location)
-        self.categories = Edit(caption='Categories: ', edit_text=self.categories)
+        self.description = ExtendedEdit(
+            caption='Description: ', edit_text=self.description, multiline=True,
+        )
+        self.location = ExtendedEdit(caption='Location: ', edit_text=self.location)
+        self.categories = ExtendedEdit(caption='Categories: ', edit_text=self.categories)
         self.alarms = AlarmsEditor(self.event)
         self.pile = NListBox(urwid.SimpleFocusListWalker([
             NColumns([self.summary, self.calendar_chooser], dividechars=2),
@@ -351,6 +376,13 @@ class EventEditor(urwid.WidgetWrap):
         ]), outermost=True)
         self._always_save = always_save
         urwid.WidgetWrap.__init__(self, self.pile)
+
+    def start_datechange(self, date):
+        self.pane.eventscolumn.original_widget.set_focus_date(date)
+        self.recurrenceeditor.update_startdt(date)
+
+    def end_datechange(self, date):
+        self.pane.eventscolumn.original_widget.set_focus_date(date)
 
     @property
     def title(self):  # Window title
@@ -434,6 +466,7 @@ class EventEditor(urwid.WidgetWrap):
             self.pane.window.alert(
                 ('light red', "Can't save: end date is before start date!"))
             return
+
         if self._always_save or self.changed is True:
             self.update_vevent()
             self.event.allday = self.startendeditor.allday
@@ -472,13 +505,16 @@ class EventEditor(urwid.WidgetWrap):
 
 WEEKDAYS = ['MO', 'TU', 'WE', 'TH', 'FR', 'SA', 'SU']  # TODO use locale and respect weekdaystart
 
+
 class WeekDaySelector(urwid.WidgetWrap):
-    def __init__(self, startdt):
+    def __init__(self, startdt, selected_days):
 
         self._weekday_boxes = {day: urwid.CheckBox(day, state=False) for day in WEEKDAYS}
-        weekday = startdt.weekday()  # TODO also needs fixing
+        weekday = startdt.weekday()
         self._weekday_boxes[WEEKDAYS[weekday]].state = True
         self.weekday_checks = NColumns([(7, self._weekday_boxes[wd]) for wd in WEEKDAYS])
+        for day in selected_days:
+            self._weekday_boxes[day].state = True
         urwid.WidgetWrap.__init__(self, self.weekday_checks)
 
     @property
@@ -490,15 +526,22 @@ class WeekDaySelector(urwid.WidgetWrap):
 class RecurrenceEditor(urwid.WidgetWrap):
 
     def __init__(self, rrule, conf, startdt):
-        # TODO: support more recurrence schemes
-        #
         self._conf = conf
-        self._startdt = startdt  # TODO make sure updated when startdt is updated in the editor
+        self._startdt = startdt
         self._rrule = rrule
         self.repeat = bool(rrule)
         self._allow_edit = not self.repeat or self.check_understood_rrule(rrule)
+        self.repeat_box = urwid.CheckBox(
+            'Repeat: ', state=self.repeat, on_state_change=self.check_repeat,
+        )
 
-        self._until = "Forever"
+        if "UNTIL" in self._rrule:
+            self._until = "Until"
+        elif "COUNT" in self._rrule:
+            self._until = "Repetitions"
+        else:
+            self._until = "Forever"
+
         recurrence = self._rrule['freq'][0].lower() if self._rrule else "weekly"
         self.recurrence_choice = Choice(
             ["daily", "weekly", "monthly", "yearly"],
@@ -509,49 +552,77 @@ class RecurrenceEditor(urwid.WidgetWrap):
             caption='every:',
             edit_text=str(self._rrule.get('INTERVAL', [1])[0]),
         )
-        self.repeat_box = urwid.CheckBox(
-            'Repeat: ', state=self.repeat, on_state_change=self.check_repeat,
-        )
         self.until_choice = Choice(
             ["Forever", "Until", "Repetitions"], self._until, callback=self.rebuild,
         )
-        self.repetitions_edit = PositiveIntEdit(edit_text='1')
-        # startdate
-        edit = ValidatedEdit(
-            dateformat=self._conf['locale']['longdateformat'],
-            EditWidget=DateWidget,
-            validate=lambda _: True,  # TODO check if valid date and in future
-            edit_text=self._startdt.strftime(self._conf['locale']['longdateformat']),
-            on_date_change=lambda _: None,
-        )
-        self.until_edit = CalendarPopUp(edit, self._conf, lambda _: None)
 
-        self.weekday_checks = WeekDaySelector(self._startdt)
+        count = str(self._rrule.get('COUNT', [1])[0])
+        self.repetitions_edit = PositiveIntEdit(edit_text=count)
 
-        xth_weekday = "every {} {}".format(
-            (self._startdt.day // 7) + 1, WEEKDAYS[self._startdt.weekday()],
-        )
-        self.monthly_choice = Choice(
-            ["on the xth", xth_weekday], "on the xth", callback=self.rebuild,
+        until = self._rrule.get('UNTIL', [None])[0]
+        if until is None and isinstance(self._startdt, datetime):
+            until = self._startdt.date()
+        elif until is None:
+            until = self._startdt
+
+        if isinstance(until, datetime):
+            until = datetime.date()
+        self.until_edit = DateEdit(
+            until, self._conf['locale']['longdateformat'],
+            lambda _: None, self._conf['locale']['weeknumbers'],
+            self._conf['locale']['firstweekday'],
         )
 
+        self._rebuild_weekday_checks()
+        self._rebuild_monthly_choice()
         self._pile = pile = NPile([urwid.Text('')])
         urwid.WidgetWrap.__init__(self, pile)
+        self.rebuild()
+
+    def _rebuild_monthly_choice(self):
+        weekday, xth = get_weekday_occurrence(self._startdt)
+        ords = {1: 'st', 2: 'nd', 3: 'rd', 21: 'st', 22: 'nd', 23: 'rd', 31: 'st'}
+        self._xth_weekday = 'on every {}{} {}'.format(
+            xth, ords.get(xth, 'th'), WEEKDAYS[weekday],
+        )
+        self._xth_monthday = 'on every {}{} of the month'.format(
+            self._startdt.day, ords.get(self._startdt.day, 'th'),
+        )
+        self.monthly_choice = Choice(
+            [self._xth_monthday, self._xth_weekday], self._xth_monthday, callback=self.rebuild,
+        )
+
+    def _rebuild_weekday_checks(self):
+        if self.recurrence_choice.active == 'weekly':
+            initial_days = self._rrule.get('BYDAY', [])
+        else:
+            initial_days = []
+        self.weekday_checks = WeekDaySelector(self._startdt, initial_days)
+
+    def update_startdt(self, startdt):
+        self._startdt = startdt
+        self._rebuild_weekday_checks()
+        self._rebuild_monthly_choice()
         self.rebuild()
 
     @staticmethod
     def check_understood_rrule(rrule):
         """test if we can reproduce `rrule`."""
         keys = set(rrule.keys())
-        unsupporetd_rrule_parts = {
-            'BYSECOND', 'BYMINUTE', 'BYHOUR', 'BYMONTHDAY', 'BYYEARDAY',
-            'BYWEEKNO', 'BYMONTH', 'BYSETPOS', 'WKST',
+        freq = rrule.get('FREQ', [None])[0]
+        unsupported_rrule_parts = {
+            'BYSECOND', 'BYMINUTE', 'BYHOUR', 'BYYEARDAY',
+            'BYWEEKNO', 'BYMONTH', 'BYSETPOS',
         }
-        if keys.intersection(unsupporetd_rrule_parts):
+        if keys.intersection(unsupported_rrule_parts):
             return False
-        if rrule.get('FREQ', [None])[0] not in ['DAILY', 'WEEKLY', 'MONTHLY', 'YEARLY']:
+        # we don't support negative BYMONTHDAY numbers
+        if rrule.get('BYMONTHDAY', ['1'])[0][0] == '-':
             return False
-        # TODO negative BYDAY
+        if freq not in ['DAILY', 'WEEKLY', 'MONTHLY', 'YEARLY']:
+            return False
+        if 'BYDAY' in keys and freq == 'YEARLY':
+            return False
         return True
 
     def check_repeat(self, checkbox, state):
@@ -575,8 +646,8 @@ class RecurrenceEditor(urwid.WidgetWrap):
             self.rebuild()
 
         pile = NPile([
-            urwid.Text("We cannot reproduce this event's repetion rules."),
-            urwid.Text("Editing the repetion rules will destroy the current rules."),
+            urwid.Text("We cannot reproduce this event's repetition rules."),
+            urwid.Text("Editing the repetition rules will destroy the current rules."),
             urwid.Button("Edit anyway", on_press=_allow_edit),
         ])
         urwid.WidgetWrap.__init__(self, pile)
@@ -613,14 +684,20 @@ class RecurrenceEditor(urwid.WidgetWrap):
         return self._rrule != self.rrule()  # TODO do this properly
 
     def rrule(self):
-        rrule = dict()   # TODO check if we need to use a caseless dict
-        rrule["freq"] = [self.recurrence_choice.active]
+        rrule = dict()
+        rrule['freq'] = [self.recurrence_choice.active]
         interval = int(self.interval_edit.get_edit_text())
         if interval != 1:
-            rrule["interval"] = [interval]
-        if rrule["freq"] == ["weekly"]:
-            if len(self.weekday_checks.days) > 1:
-                rrule["byday"] = self.weekday_checks.days
+            rrule['interval'] = [interval]
+        if rrule['freq'] == ['weekly'] and len(self.weekday_checks.days) > 1:
+            rrule['byday'] = self.weekday_checks.days
+        if rrule['freq'] == ['monthly'] and self.monthly_choice.active == self._xth_weekday:
+            weekday, occurrence = get_weekday_occurrence(self._startdt)
+            rrule['byday'] = ['{}{}'.format(occurrence, WEEKDAYS[weekday])]
+        if self.until_choice.active == 'Until':
+            rrule['until'] = self.until_edit.date
+        elif self.until_choice.active == 'Repetitions':
+            rrule['count'] = int(self.repetitions_edit.get_edit_text())
         return rrule
 
     @property
@@ -641,7 +718,7 @@ class ExportDialog(urwid.WidgetWrap):
         lines = []
         lines.append(urwid.Text('Export event as ICS file'))
         lines.append(urwid.Text(''))
-        export_location = Edit(
+        export_location = ExtendedEdit(
             caption='Location: ', edit_text="~/%s.ics" % event.summary.strip())
         lines.append(export_location)
         lines.append(urwid.Divider(' '))